+++ src/pages/admin/GiftCards.vue
@@ -1,20 +1,23 @@
 <template>
   <div>
-    <et-grid type="giftcard" :disableEditting="true" :expandedCol="expandedCol" />
+    <et-grid type="giftcard" :disableEditting="true" :expandedCol="expandedCol" supportGiftCardsPrinting />
     <modal-create-gift-card />
     <modal-q-r-code :qrcode="getCurrentGenGiftCardCode" />
+    <gift-cards-printing>
   </div>
 </template>
 <script>
 import modalCreateGiftCard from '../../components/CreateGiftCardModal'
 import modalQRCode from '../../components/QRCodeModal'
 import etGrid from '../../components/EtGrid'
+import giftCardsPrinting from '../../components/GiftCardsPrinting'
 import {mapGetters, mapMutations} from 'vuex'
 export default {
   components: {
     etGrid,
     modalQRCode,
     modalCreateGiftCard,
+    giftCardsPrinting,
   },
   computed: {
     ...mapGetters('giftcard', ['getCurrentGenGiftCardCode', 'getIsModalQRCodeOpened']),